pyproject.toml
[tool.poetry]
name = "cultura-project"
version = "0.1.0"
description = ""
authors = ["Charles De Dampierre <charlesdedampierre@gmail.com>"]
readme = "README.md"
packages = [{ include = "cultura_project" }]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
python-dotenv = "^1.0.0"
bambi = "^0.13.0"
arviz = "^0.16.1"
statsmodels = "^0.14.0"
plotly = "^5.18.0"
pio = "^0.0.3"
geopandas = "^0.14.1"
seaborn = "^0.13.0"
matplotlib-venn = "^0.11.9"
scikit-learn = "^1.3.2"
ipyleaflet = "^0.18.0"
selenium = "^4.16.0"
networkx = "^3.2.1"
sparqlwrapper = "^2.0.0"
pydantic = "^2.5.3"
jax = "^0.4.23"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
